siteauto/requirements.txt
asgiref==3.9.2
asttokens==3.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.7
django-debug-toolbar==6.1.0
django-extensions==4.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.3
executing==2.2.1
git-filter-repo==2.47.0
hiredis==3.3.0
idna==3.11
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
oauthlib==3.3.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
python3-openid==3.2.0
psycopg2
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.6.0
wcwidth==0.2.14
gevent==24.11.1
gunicorn==23.0.0
